backend/requirements.txt
google-cloud-documentai
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-storage
pydantic
pytest
pytest-asyncio
pytest-mock
